[Refactor] package 정리
