+++ src/main/kotlin/exception/DivisibleMoneyException.kt
@@ -1,3 +1,3 @@
-package lotto.exception
+package exception
 
 class DivisibleMoneyException : Exception()
