+++ src/main/kotlin/lotto/Lotto.kt
@@ -1,11 +1,12 @@
 package lotto
 
-import lotto.constant.LOTTO_NUM_COUNT
-import lotto.constant.LOTTO_NUM_RANGE_END
-import lotto.exception.DuplicateInputException
-import lotto.util.ErrorType
-import lotto.util.printErrorMessage
-import lotto.util.showError
+import constant.LOTTO_NUM_COUNT
+import constant.LOTTO_NUM_RANGE_END
+import exception.DuplicateInputException
+import util.ErrorType
+import util.printErrorMessage
+import util.showError
+import winlotto.WinLotto
 
 class Lotto(private val numbers: List<Int>) {
     init {
