+++ src/main/kotlin/winlotto/WinLotto.kt
@@ -1,12 +1,12 @@
-package lotto
+package winlotto
 
-import lotto.constant.LOTTO_BONUS_NUM_COUNT
-import lotto.constant.LOTTO_NUM_COUNT
-import lotto.constant.LOTTO_NUM_RANGE_END
-import lotto.exception.DuplicateInputException
-import lotto.util.ErrorType
-import lotto.util.printErrorMessage
-import lotto.util.showError
+import constant.LOTTO_BONUS_NUM_COUNT
+import constant.LOTTO_NUM_COUNT
+import constant.LOTTO_NUM_RANGE_END
+import exception.DuplicateInputException
+import util.ErrorType
+import util.printErrorMessage
+import util.showError
 
 class WinLotto(val numbers: List<Int>) {
     init {
