+++ src/main/kotlin/winlotto/WinLottoFactory.kt
@@ -1,4 +1,4 @@
-package lotto
+package winlotto
 
 class WinLottoFactory {
     fun makeWinLotto(
